requirements.txt

fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
openai>=1.6.0
google-generativeai>=0.3.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
httpx>=0.25.0
docker>=6.0.0
pytest>=7.0.0